updated status and note column
To automatically check for raw data availability the status column
needs to comply with the defined rules for data availability status
(see README.md). Also, the notes column is necessary for separating
special cases.

The # character in column 1 indicates that the row should be ignored
unless the status is changed to Available.